+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/Source.java
@@ -438,6 +438,7 @@ public final class Source {
         private String name;
         private boolean interactive;
         private boolean internal;
+        private boolean cached = true;
         private CharSequence content;
 
         Builder(String language, Object origin) {
@@ -522,6 +523,25 @@ public final class Source {
         }
 
         /**
+         * Enables or disables code caching for this source. By default code caching is enabled. If
+         * <code>true</code> then the source does not require parsing every time this source is
+         * {@link Context#eval(Source) evaluated}. If <code>false</code> then the source requires
+         * parsing every time the source is evaluated but does not remember any state. Disabling
+         * caching may be useful if the source is known to only be evaluated once.
+         * <p>
+         * If a source instance is no longer referenced by the client then all code caches will be
+         * freed automatically. Also, if the underlying context or engine is no longer referenced
+         * then cached code for evaluated sources will be freed automatically.
+         *
+         * @return instance of <code>this</code> builder ready to {@link #build() create new source}
+         * @since 1.0
+         */
+        public Builder cached(@SuppressWarnings("hiding") boolean cached) {
+            this.cached = cached;
+            return this;
+        }
+
+        /**
          * Assigns new {@link URI} to the {@link #build() to-be-created} {@link Source}. Each source
          * provides {@link Source#getURI()} as a persistent identification of its location. A
          * default value for the method is deduced from the location or content, but one can change
@@ -545,7 +565,7 @@ public final class Source {
          * @since 1.0
          */
         public Source build() throws IOException {
-            return getImpl().build(language, origin, uri, name, content, interactive, internal);
+            return getImpl().build(language, origin, uri, name, content, interactive, internal, cached);
         }
 
         /**